+++ includes/Helpers/Cache.php
@@ -2,8 +2,10 @@
 
 namespace Tpay\Helpers;
 
-class Cache
+class Cache extends \Tpay\OpenApi\Utilities\Cache
 {
+    public function __construct() {}
+
     public function get($key)
     {
         $key .= ':';
@@ -17,6 +19,13 @@ class Cache
         $key .= ':';
         $key .= tpayOption('global_tpay_environment');
         set_transient($key, $value, $ttl);
+    }
+
+    public function delete($key)
+    {
+        $key .= ':';
+        $key .= tpayOption('global_tpay_environment');
+        delete_transient($key);
     }
 
     public function erase(): void