+++ includes/Helpers/CardHelper.php
@@ -8,8 +8,12 @@ class CardHelper
 
     public function save_card(array $card): bool
     {
-        $hash = sha1($card['card_hash'].WP_TPAY_HASH);
-        DatabaseConnection::delete(self::TABLE, ['hash' => $hash, 'source_order' => null]);
+        if (defined('WP_TPAY_HASH')) {
+            $hash = sha1($card['card_hash'].WP_TPAY_HASH);
+            DatabaseConnection::delete(self::TABLE, ['hash' => $hash, 'source_order' => null]);
+        } else {
+            $hash = md5($card['card_hash']);
+        }
         $card_id = DatabaseConnection::queryVar('SELECT id FROM %i WHERE hash = %s', self::TABLE, $hash);
 
         if (!$card_id) {